[REFACTOR] deposit 분리
