--- deposit/src/main/java/io/devground/common/aop/GlobalExceptionLoggingAspect.java
@@ -1,20 +1,0 @@
-package io.devground.common.aop;
-
-import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.annotation.AfterThrowing;
-import org.aspectj.lang.annotation.Aspect;
-import org.springframework.stereotype.Component;
-
-import io.devground.common.util.LogUtil;
-import lombok.RequiredArgsConstructor;
-
-@Aspect
-@Component
-@RequiredArgsConstructor
-public class GlobalExceptionLoggingAspect {
-
-	@AfterThrowing(pointcut = "@within(org.springframework.stereotype.Service)", throwing = "ex")
-	public void logGlobalException(JoinPoint joinPoint, Throwable ex) {
-		LogUtil.logError(joinPoint, ex);
-	}
-}
