--- deposit/src/main/java/io/devground/common/kafka/KafkaTemplateConfig.java
@@ -1,17 +1,0 @@
-package io.devground.common.kafka;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-
-@Configuration
-public class KafkaTemplateConfig {
-
-	@Bean
-	public KafkaTemplate<String, Object> kafkaTemplate(
-		ProducerFactory<String, Object> producerFactory
-	) {
-		return new KafkaTemplate<>(producerFactory);
-	}
-}
