--- deposit/src/main/java/io/devground/common/saga/vo/SagaStatus.java
@@ -1,25 +1,0 @@
-package io.devground.common.saga.vo;
-
-import lombok.Getter;
-
-@Getter
-public enum SagaStatus {
-
-	IN_PROCESS,
-	SUCCESS,
-	FAILED,
-	COMPENSATING,
-	COMPENSATED;
-
-	public boolean isProcess() {
-		return this == IN_PROCESS;
-	}
-
-	public boolean isTerminal() {
-		return this == SUCCESS || this == FAILED || this == COMPENSATED;
-	}
-
-	public boolean isCompensating() {
-		return this == COMPENSATING;
-	}
-}
