--- deposit/src/main/java/io/devground/common/saga/vo/SagaType.java
@@ -1,14 +1,0 @@
-package io.devground.common.saga.vo;
-
-import lombok.Getter;
-
-@Getter
-public enum SagaType {
-
-	PRODUCT_IMAGE_REGIST,
-	PRODUCT_IMAGE_UPDATE,
-	PRODUCT_IMAGE_DELETE,
-
-	// Settlement
-	SETTLEMENT_DEPOSIT_CHARGE
-}
