+++ deposit/src/main/java/io/devground/dbay/DepositApplication.java
@@ -1,0 +1,13 @@
+package io.devground.dbay;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DepositApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DepositApplication.class, args);
+	}
+
+}
