+++ deposit/src/main/java/io/devground/dbay/common/aop/BaseResponseAspect.java
@@ -1,0 +1,37 @@
+package io.devground.dbay.common.aop;
+
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.springframework.stereotype.Component;
+
+import io.devground.core.model.web.BaseResponse;
+import io.devground.dbay.common.util.LogUtil;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+
+@Aspect
+@Component
+@RequiredArgsConstructor
+public class BaseResponseAspect {
+
+	private final HttpServletResponse response;
+
+	@Around("within(@org.springframework.web.bind.annotation.RestController *)")
+	public Object updateHttpStatus(ProceedingJoinPoint joinPoint) throws Throwable {
+		String className = joinPoint.getSignature().getDeclaringType().getSimpleName();
+		String methodName = joinPoint.getSignature().getName();
+
+		LogUtil.logControllerRequest(className, methodName);
+
+		Object proceed = joinPoint.proceed();
+
+		if (proceed instanceof BaseResponse<?> baseResponse) {
+			LogUtil.logControllerResponse(className, methodName, baseResponse);
+
+			response.setStatus(baseResponse.resultCode());
+		}
+
+		return proceed;
+	}
+}
