+++ deposit/src/main/java/io/devground/dbay/common/kafka/KafkaTemplateConfig.java
@@ -1,0 +1,17 @@
+package io.devground.dbay.common.kafka;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.core.ProducerFactory;
+
+@Configuration
+public class KafkaTemplateConfig {
+
+	@Bean
+	public KafkaTemplate<String, Object> kafkaTemplate(
+		ProducerFactory<String, Object> producerFactory
+	) {
+		return new KafkaTemplate<>(producerFactory);
+	}
+}
