+++ deposit/src/main/java/io/devground/dbay/common/saga/vo/SagaType.java
@@ -1,0 +1,14 @@
+package io.devground.dbay.common.saga.vo;
+
+import lombok.Getter;
+
+@Getter
+public enum SagaType {
+
+	PRODUCT_IMAGE_REGIST,
+	PRODUCT_IMAGE_UPDATE,
+	PRODUCT_IMAGE_DELETE,
+
+	// Settlement
+	SETTLEMENT_DEPOSIT_CHARGE
+}
