--- deposit/src/main/java/io/devground/deposit/DepositApplication.java
@@ -1,13 +1,0 @@
-package io.devground.deposit;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DepositApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DepositApplication.class, args);
-	}
-
-}
