--- deposit/src/main/java/io/devground/deposit/repository/DepositJpaRepository.java
@@ -1,18 +1,0 @@
-package io.devground.deposit.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import io.devground.deposit.entity.Deposit;
-
-public interface DepositJpaRepository extends JpaRepository<Deposit, Long> {
-
-	Optional<Deposit> findByUserCode(String userCode);
-
-	boolean existsByUserCode(String userCode);
-
-	Optional<Deposit> findByCode(String code);
-
-	void deleteByUserCode(String userCode);
-}
